% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDNA_power.R
\name{eDNA_power}
\alias{eDNA_power}
\title{Compute Power Analysis for eDNA Hierarchical Detection Sampling Design}
\usage{
eDNA_power(n, k, theta, p, s_min, n_sim = 10000)
}
\arguments{
\item{n}{Integer. Number of samples collected.}

\item{k}{Integer. Number of technical replicates per sample.}

\item{theta}{Numeric or Numeric Vector. Value(s) between 0 and 1. Probability that DNA is present in a sample.}

\item{p}{Numeric or Numeric Vector. Value(s) between 0 and 1. Probability that a technical replicate detects the DNA if it is present.}

\item{s_min}{Integer. Minimum number of positive samples to be considered successful.}

\item{n_sim}{Integer. Number of Monte Carlo simulations to run (default 10000). Only used if a vector is provided for `theta` or `p`.}
}
\value{
Numeric. The probability of ≥ s_min positive samples, considering at least one technical replicate per sample detecting DNA as a positive detection.
}
\description{
Computes the probability of observing at least `s_min` positive samples
given the number of samples, number of technical replicates, and values or distributions for detection probabilities.
This function considers at least one technical replicate per sample detecting DNA as a positive detection.
}
\examples{
eDNA_power(n = 10, k = 3, theta = 0.6, p = 0.8, s_min = 5)
eDNA_power(n = 15, k = 2, theta = runif(1000, 0.4, 0.6), p = rbeta(1000, 1.8, 4.2), s_min = 5)
}
