% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDNA_power.R
\name{eDNA_power_extract}
\alias{eDNA_power_extract}
\title{Extract n sample and k replicates that acheive target power}
\usage{
eDNA_power_extract(grid, target_power)
}
\arguments{
\item{grid}{Data frame returned by `eDNA_power_grid`.}

\item{target_power}{Numeric. Target power level between 0 and 1.}
}
\value{
Data frame with minimum `n` and `k` to achieve target power.
}
\description{
Extracts the minimum number of samples (n) and technical replicates (k) needed to achieve a target power level.
}
\examples{
grid <- eDNA_power_grid(n_vals = 5:20, k_vals = 1:5, theta = 0.6, p = 0.8, s_min = 5)
eDNA_power_extract(grid)


}
