% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_power.R
\name{freq_power_grid}
\alias{freq_power_grid}
\title{Compute Frequentist Power Grid}
\usage{
freq_power_grid(n_vals, k_vals, theta, p, s_min = 5)
}
\arguments{
\item{n_vals}{Integer vector. Possible number of samples.}

\item{k_vals}{Integer vector. Possible numbers of technical replicates per sample.}

\item{theta}{Numeric between 0 and 1. Probability species is present in a sample.}

\item{p}{Numeric between 0 and 1. Probability a technical replicate detects the species if present.}

\item{s_min}{Integer. Minimum number of positive samples to be considered successful (default 5).}
}
\value{
Data frame with columns `n`, `k`, and `power`.
}
\description{
Computes frequentist power across a grid of sample sizes (`n_vals`) and technical replicates (`k_vals`).
}
\examples{
freq_power_grid(5:10, 1:3, theta = 0.6, p = 0.8)
}
