% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_power.R
\name{plot_bayes_heatmap}
\alias{plot_bayes_heatmap}
\title{Plot Bayesian Power Heatmap}
\usage{
plot_bayes_heatmap(grid)
}
\arguments{
\item{grid}{Data frame returned by `bayes_power_grid`.}
}
\value{
ggplot2 object displaying a heatmap.
}
\description{
Generates a heatmap of Bayesian prior predictive power across combinations of `n` and `k`.
}
\examples{
grid <- bayes_power_grid(5:10, 1:3, s_min = 5, a_theta = 6, b_theta = 4, a_p = 8, b_p = 2)
plot_bayes_heatmap(grid)
}
