% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eDNA_power.R
\name{eDNA_power_grid}
\alias{eDNA_power_grid}
\title{Compute Power Grid}
\usage{
eDNA_power_grid(n_vals, k_vals, s_min, n_sim = 10000, theta = NULL, p = NULL)
}
\arguments{
\item{n_vals}{Integer vector. Possible number of samples.}

\item{k_vals}{Integer vector. Possible numbers of technical replicates per sample.}

\item{s_min}{Integer. Minimum number of positive samples to be considered successful.}

\item{n_sim}{Integer. Number of Monte Carlo simulations to run (default 10000). Only used if a vector is provided for `theta` or `p`.}

\item{theta}{Numeric or Numeric Vector. Value(s) between 0 and 1. Probability that DNA is present in a sample.}

\item{p}{Numeric or  Numeric Vector. Value(s) between 0 and 1. Probability that a technical replicate detects the DNA if it is present.}
}
\value{
Data frame with columns `n`, `k`, and `power`.
}
\description{
Computes prior predictive power across a grid of sample sizes (`n_vals`) and technical replicates (`k_vals`).
}
\examples{
eDNA_power_grid(n_vals = 5:10, k_vals = 1:3, theta = 0.6, p = 0.8, s_min = 5)
eDNA_power_grid(n_vals = 5:10, k_vals = 1:3, theta = runif(1000, 0.4, 0.6), p = rbeta(1000, 1.8, 4.2), s_min = 5)
}
